+++ src/rootcheck/rootcheck.c
@@ -170,13 +170,15 @@ int rootcheck_init(int test_config)
         mterror_exit(ARGV0, CONFIG_ERROR, cfg);
     }
 
+#ifndef WIN32
     if(rootcheck.checks.rc_unixaudit) {
-        mwarn("The check_unixaudit option is deprecated in favor of the Configuration Assessment module.");
+        mwarn("The check_unixaudit option is deprecated in favor of the SCA module.");
     }
+#endif
 
 #ifdef WIN32
     if(rootcheck.checks.rc_winaudit) {
-        mwarn("The check_winaudit option is deprecated in favor of the Configuration Assessment module.");
+        mwarn("The check_winaudit option is deprecated in favor of the SCA module.");
     }
 #endif